experiments/aot: re-measure on v1.13.0 and correct the encoder's story (#40)
* experiments/aot: re-measure on tip, and correct the encoder's story

The Result table paired a v1.11.0 VM baseline with a v1.11.0 native port
measurement, and the ~83x encode figure derived from it has been quoted
elsewhere since. Re-measured everything on tip bdd8268c (v1.12.2-95), one
machine, same day: the native port's sixel encode is 1.635 ms, not ~1.4 ms, and
the VM's is 151 ms, not ~120 ms. Encode is 92x, not 83x.

The bigger correction is qualitative. This table said the encoder "can't lower
today" — it does lower now, but into Go that keeps every element access on a
boxed ec.Invoke, so it buys 2.2-2.6x rather than nothing. That reframes #358:
the prize is no longer the whole ~83x, it is the 35-42x still sitting between
the AOT output and the hand-written port. Numbers came from building the demo
through lg-compile --entry-frame, which is a path that did not exist when this
was written.

Adds an AOT column, since VM-versus-native alone no longer describes what the
toolchain does. Two caveats the column needs: the compute figure requires
^double hints, because the demo's own unhinted escape lowers to int params and
the call site's type guard silently falls back to the VM (#551); and building
the demo this way needs three math/* sites hand-unboxed to compile (#562), with
the resulting binary running -main twice.

The old v1.11.0 figures stay in Caveats as provenance, labelled as
unreconciled, so the next person has the pairing they came from rather than a
loose number to re-mix.


* experiments/aot: re-measure on v1.13.0, and follow int -> int64

lg-compile's lowered signatures widened int to int64 in v1.13.0 (the 64-bit-
across-the-rt-boundary work, nooga/let-go#906). Nothing warns; the Go that calls
the generated package just stops compiling. Two callers here did:

  - aot_bench_test.go assigned the result to an int. Fixed with an explicit
    int64() conversion, which is a no-op on a build that already returns int64
    and widens on an older one, so the template still compiles against both.
  - demos/mandelbrot-aot's hand-written driver, in seven places across the
    render, checksum and bench paths. Argument position has no both-ways
    spelling, so those convert at the call site and target the current
    signatures.

go.mod follows to Go 1.27, which v1.13.0 requires.

Re-measured with those building again. The Result table above is deliberately
NOT rewritten: its AOT column needs the manual #562 hand-unboxing to reproduce,
and replacing only the other three columns would produce the mixed-vintage table
the notes already warn against. Today's comparable numbers go in a dated note
instead -- VM 115/151/266 ms -> 96/129/225, native 1.348/1.635/3.038 ->
1.244/1.495/2.971, AOT micro-kernel 1.343 -> 1.268. Workload unchanged
(bytes=27695).

The VM gained 15-17% against the port's 2-8%, so the ratios narrow rather than
the port regressing. Also noted: #902 (entry-frame binaries running -main twice)
is fixed in v1.13.0, and that short benchtimes on a loaded box lie -- a 2s pass
put the AOT kernel 24% high and three 3s passes did not reproduce it.


* experiments/aot: route the repro scripts through the lg resolver, raise the floor to 1.13

Review findings on #40.

[P1] Both documented reproduction paths failed before compiling anything. #39
moved lg resolution into scripts/lib/lg-path.sh, but build.sh and regen-aot.sh
still spelled the binary as the checkout root ./lg, which a current let-go does
not produce — it builds to build/lg and promotes to bin/lg. On a v1.13.0
checkout that is "./lg: No such file or directory" and "no lg at .../lg". Both
now source the resolver. build.sh keeps $LG as the CHECKOUT, since lg-compile
has to run from inside it, and adds $LGBIN for the binary.

[P2] The advertised >= 1.12 floor no longer builds. The int64 call sites target
v1.13.0 signatures (#906), so on 1.12 the generated kernel lowers to int and
native/main.go fails with eight type errors in a file the user did not write.
Keeping 1.12 was not available: the return value converts both ways, but
argument position has no spelling that compiles against int and int64 at once.
So the floor is raised rather than faked. README, go.mod and build.sh's header
now say 1.13, and the signature example shows the int64 form with a note that it
was int before v1.13.0.

build.sh also gains a second version guard beside the existing pre-1.12 one,
reading the generated signature rather than any version string, so 1.12 reports
what is wrong instead of leaving Go to report it from generated code.

Verified on a checkout with bin/lg and no root lg: build.sh builds and
regen-aot.sh benches (1.62 ms). Verified on a real v1.12.2 build: the new guard
fires with its message, and no Go type errors are reached.


---------

+++ demos/mandelbrot-aot/native/main.go
@@ -46,8 +46,8 @@ func render(W, H int, cx0, cy0, spanX float64, mi int) []byte {
 		cy := cy0 - spanY/2 + spanY*float64(row)/float64(H)
 		for col := 0; col < W; col++ {
 			cx := cx0 - spanX/2 + spanX*float64(col)/float64(W)
-			it := k.Escape(nil, cx, cy, mi)
-			if it >= mi { // interior
+			it := k.Escape(nil, cx, cy, int64(mi))
+			if it >= int64(mi) { // interior
 				if colorOn {
 					buf = append(buf, "\x1b[38;2;10;10;20m "...)
 				} else {
@@ -60,7 +60,7 @@ func render(W, H int, cx0, cy0, spanX float64, mi int) []byte {
 				buf = appendCell(buf, int(9+246*t), int(20+120*t*t), int(120+135*(1-t)), ramp[int(t*float64(len(ramp)-1))])
 			} else {
 				// mi-independent: char from raw iteration count → no maxiter coupling.
-				buf = append(buf, ramp[it%len(ramp)])
+				buf = append(buf, ramp[int(it)%len(ramp)])
 			}
 		}
 		if colorOn {
@@ -332,9 +332,9 @@ func main() {
 				cy := cy0 - spanY/2 + spanY*float64(row)/float64(H)
 				for col := 0; col < W; col++ {
 					cx := cx0 - span/2 + span*float64(col)/float64(W)
-					it := k.Escape(nil, cx, cy, mi)
-					sum += int64(it)
-					if it >= mi {
+					it := k.Escape(nil, cx, cy, int64(mi))
+					sum += it
+					if it >= int64(mi) {
 						interior++
 					} else if row < H/2 && col < W/2 {
 						ulDetail++ // escaping (non-black) cells in the upper-left quadrant
@@ -362,9 +362,9 @@ func main() {
 		for r := 0; r < reps; r++ {
 			sum = 0
 			for row := 0; row < H; row++ {
-				cy := k.CyOf(nil, row)
+				cy := k.CyOf(nil, int64(row))
 				for col := 0; col < W; col++ {
-					sum += int64(k.Escape(nil, k.CxOf(nil, col), cy, MI))
+					sum += k.Escape(nil, k.CxOf(nil, int64(col)), cy, MI)
 				}
 			}
 		}
